pyproject.toml
[tool.poetry]
name = "ibis-extensions"
version = "0.1.2"
description = ""
authors = ["Omri Fima"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
ibis-framework = ">=4.1.0,<5.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
ibis-framework = {extras = ["duckdb"], version = ">=4.1.0"}
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"